+++ Amazing Numbers/task/src/numbers/DuckNumberDeterminer.java
@@ -3,8 +3,8 @@ package numbers;
 public class DuckNumberDeterminer implements NumbersDeterminer {
 
     @Override
-    public boolean getProperty(int number) {
-        for (char ch : Integer.toString(number).toCharArray()) {
+    public boolean getProperty(long number) {
+        for (char ch : Long.toString(number).toCharArray()) {
             if (ch == '0') return true;
         }
         return false;
